refactor: UIバックエンドInstallerの共通ロジックをInstallerBaseへ抽出
ImmediateModeInstaller(683行)とUIToolkitInstaller(842行)に、コメント・インデントを
除いて完全に同一のコピーが約300行あった。uGUI版を追加すると3つ目の複製になるため、
抽象基底クラスへ集約する(#137)。

- InstallerBase を追加し、以下を集約
  - Install/Uninstall/UninstallAsync/Resolve と失敗時のCleanUpフロー
  - Domain構築・コマンド登録・入力生成・Coordinator構築・Scope構築
  - _theme/_animation/_options の[SerializeReference]フィールドとnull時のフォールバック
  - DomainContext/RenderingContext/CoordinatorContext を protected 入れ子型へ昇格
- 派生に残す拡張点は BuildRenderingContext/ClearReferences/SyncTheme のみ
- UIToolkit固有の関心事を分離
  - UIToolkitViewFactory: GameObject/UIDocument/WindowRootの生成とアセット解決
  - UIToolkitThemeApplier: VisualElementへのテーマ色・フォント適用
  - RuntimeGeneratedAsset: 実行時生成PanelSettingsの解放ハンドル。Componentsへ載せて
    TerminalRuntimeScopeに破棄させる

ImmediateModeInstaller 683→212行、UIToolkitInstaller 842→227行。
具象Installerの型名・名前空間・アセンブリは不変のため、既存シーンのSerializeReference
参照は壊れない。
