+++ Assets/YukimaruGames/Terminal/Runtime/Composition/Installer/InstallerBase.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 0d0ff1b85123f4ccbb94721e964fb63c
