+++ Assets/YukimaruGames/Terminal/Runtime/Composition/Installer/UIToolkitViewFactory.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 0022f96d84ba144ffb68be5ee81eeb90
